+++ experiments/tune/cleartext2_tune/conf/data/benchls.yaml
@@ -1,4 +1,8 @@
-_target_: tune.BenchLSDataModule
+_target_: cleartext2_tune.data.BenchLSDataModule
+tokenizer:
+  _target_: transformers.AutoTokenizer.from_pretrained
+  pretrained_model_name_or_path: ${model_name}
+  clean_up_tokenization_spaces: True
 path: ${data_dir}/BenchLS.txt
 train_val_split: ${train_val_split}
 batch_size: ${batch_size}
