+++ experiments/tune/cleartext2_tune/conf/tune.yaml
@@ -19,4 +19,5 @@ top_k: 10
 mask: True
 
 defaults:
-  - data: benchls.yaml
+  - data: benchls
+  - module: ls
